Split tray app into its own crate
- Add `playitd-tray` as a workspace member
- Move Windows tray-specific dependencies into the new package
- Remove the tray binary and related deps from `playitd`
