+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "packages/playit-cli",
     "packages/playit-ipc",
     "packages/playitd",
+    "packages/playitd-tray",
     "packages/agent_core",
     "packages/agent_proto",
     "packages/api_client",
