+++ packages/agent_proto/src/udp_proto.rs
@@ -80,14 +80,18 @@ impl UdpFlow {
     pub fn src(&self) -> SocketAddr {
         match self {
             UdpFlow::V4 { src, .. } => SocketAddr::V4(*src),
-            UdpFlow::V6 { src: (ip, port), .. } => SocketAddr::V6(SocketAddrV6::new(*ip, *port, 0, 0)),
+            UdpFlow::V6 {
+                src: (ip, port), ..
+            } => SocketAddr::V6(SocketAddrV6::new(*ip, *port, 0, 0)),
         }
     }
 
     pub fn dst(&self) -> SocketAddr {
         match self {
             UdpFlow::V4 { dst, .. } => SocketAddr::V4(*dst),
-            UdpFlow::V6 { dst: (ip, port), .. } => SocketAddr::V6(SocketAddrV6::new(*ip, *port, 0, 0)),
+            UdpFlow::V6 {
+                dst: (ip, port), ..
+            } => SocketAddr::V6(SocketAddrV6::new(*ip, *port, 0, 0)),
         }
     }
 
@@ -97,15 +101,24 @@ impl UdpFlow {
         }
 
         match self {
-            UdpFlow::V4 { src, dst, frag, extension } => {
+            UdpFlow::V4 {
+                src,
+                dst,
+                frag,
+                extension,
+            } => {
                 slice.write_u32::<BigEndian>((*src.ip()).into()).unwrap();
                 slice.write_u32::<BigEndian>((*dst.ip()).into()).unwrap();
                 slice.write_u16::<BigEndian>(src.port()).unwrap();
                 slice.write_u16::<BigEndian>(dst.port()).unwrap();
 
                 if let Some(extension) = extension {
-                    slice.write_u64::<BigEndian>(extension.client_server_id.get()).unwrap();
-                    slice.write_u64::<BigEndian>(extension.tunnel_id.get()).unwrap();
+                    slice
+                        .write_u64::<BigEndian>(extension.client_server_id.get())
+                        .unwrap();
+                    slice
+                        .write_u64::<BigEndian>(extension.tunnel_id.get())
+                        .unwrap();
                     slice.write_u16::<BigEndian>(extension.port_offset).unwrap();
 
                     match frag {
@@ -120,26 +133,42 @@ impl UdpFlow {
                         }
                     }
 
-                    slice.write_u64::<BigEndian>(REDIRECT_FLOW_4_FOOTER_ID_V2).unwrap();
+                    slice
+                        .write_u64::<BigEndian>(REDIRECT_FLOW_4_FOOTER_ID_V2)
+                        .unwrap();
                 } else {
-                    slice.write_u64::<BigEndian>(REDIRECT_FLOW_4_FOOTER_ID_V1).unwrap()
+                    slice
+                        .write_u64::<BigEndian>(REDIRECT_FLOW_4_FOOTER_ID_V1)
+                        .unwrap()
                 }
             }
-            UdpFlow::V6 { src, dst, extension } => {
+            UdpFlow::V6 {
+                src,
+                dst,
+                extension,
+            } => {
                 slice.write_u128::<BigEndian>(src.0.into()).unwrap();
                 slice.write_u128::<BigEndian>(dst.0.into()).unwrap();
                 slice.write_u16::<BigEndian>(src.1).unwrap();
                 slice.write_u16::<BigEndian>(dst.1).unwrap();
 
                 if let Some(extension) = extension {
-                    slice.write_u64::<BigEndian>(extension.client_server_id.get()).unwrap();
-                    slice.write_u64::<BigEndian>(extension.tunnel_id.get()).unwrap();
+                    slice
+                        .write_u64::<BigEndian>(extension.client_server_id.get())
+                        .unwrap();
+                    slice
+                        .write_u64::<BigEndian>(extension.tunnel_id.get())
+                        .unwrap();
                     slice.write_u16::<BigEndian>(extension.port_offset).unwrap();
-                    slice.write_u64::<BigEndian>(REDIRECT_FLOW_6_FOOTER_ID_V2).unwrap();
+                    slice
+                        .write_u64::<BigEndian>(REDIRECT_FLOW_6_FOOTER_ID_V2)
+                        .unwrap();
                 } else {
                     /* flow label (no longer used) */
                     slice.write_u32::<BigEndian>(0).unwrap();
-                    slice.write_u64::<BigEndian>(REDIRECT_FLOW_6_FOOTER_ID_V1).unwrap();
+                    slice
+                        .write_u64::<BigEndian>(REDIRECT_FLOW_6_FOOTER_ID_V1)
+                        .unwrap();
                 }
             }
         }
@@ -197,15 +226,21 @@ impl UdpFlow {
                 let dst_ip = slice.read_u32::<BigEndian>().unwrap();
                 let src_port = slice.read_u16::<BigEndian>().unwrap();
                 let dst_port = slice.read_u16::<BigEndian>().unwrap();
-                let client_server_id = NonZeroU64::new(slice.read_u64::<BigEndian>().unwrap()).ok_or(None)?;
-                let tunnel_id = NonZeroU64::new(slice.read_u64::<BigEndian>().unwrap()).ok_or(None)?;
+                let client_server_id =
+                    NonZeroU64::new(slice.read_u64::<BigEndian>().unwrap()).ok_or(None)?;
+                let tunnel_id =
+                    NonZeroU64::new(slice.read_u64::<BigEndian>().unwrap()).ok_or(None)?;
                 let port_offset = slice.read_u16::<BigEndian>().unwrap();
 
                 let frag = if let Some(packet_id) = NonZeroU16::new(packet_id) {
                     let has_more = slice.read_u8().unwrap() != 0;
                     let frag_offset = slice.read_u16::<BigEndian>().unwrap();
 
-                    Some(FragmentInfo { packet_id, frag_offset, has_more })
+                    Some(FragmentInfo {
+                        packet_id,
+                        frag_offset,
+                        has_more,
+                    })
                 } else {
                     None
                 };
@@ -214,7 +249,11 @@ impl UdpFlow {
                     src: SocketAddrV4::new(src_ip.into(), src_port),
                     dst: SocketAddrV4::new(dst_ip.into(), dst_port),
                     frag,
-                    extension: Some(UdpFlowExtension { client_server_id, tunnel_id, port_offset }),
+                    extension: Some(UdpFlowExtension {
+                        client_server_id,
+                        tunnel_id,
+                        port_offset,
+                    }),
                 })
             }
             REDIRECT_FLOW_6_FOOTER_ID_V1 => {
@@ -230,7 +269,11 @@ impl UdpFlow {
                 let dst_port = slice.read_u16::<BigEndian>().unwrap();
                 let _flow = slice.read_u32::<BigEndian>().unwrap();
 
-                Ok(UdpFlow::V6 { src: (src_ip.into(), src_port), dst: (dst_ip.into(), dst_port), extension: None })
+                Ok(UdpFlow::V6 {
+                    src: (src_ip.into(), src_port),
+                    dst: (dst_ip.into(), dst_port),
+                    extension: None,
+                })
             }
             REDIRECT_FLOW_6_FOOTER_ID_V2 => {
                 if slice.len() < IP6_LEN_V2 {
@@ -244,14 +287,20 @@ impl UdpFlow {
                 let src_port = slice.read_u16::<BigEndian>().unwrap();
                 let dst_port = slice.read_u16::<BigEndian>().unwrap();
 
-                let client_server_id = NonZeroU64::new(slice.read_u64::<BigEndian>().unwrap()).ok_or(None)?;
-                let tunnel_id = NonZeroU64::new(slice.read_u64::<BigEndian>().unwrap()).ok_or(None)?;
+                let client_server_id =
+                    NonZeroU64::new(slice.read_u64::<BigEndian>().unwrap()).ok_or(None)?;
+                let tunnel_id =
+                    NonZeroU64::new(slice.read_u64::<BigEndian>().unwrap()).ok_or(None)?;
                 let port_offset = slice.read_u16::<BigEndian>().unwrap();
 
                 Ok(UdpFlow::V6 {
                     src: (src_ip.into(), src_port),
                     dst: (dst_ip.into(), dst_port),
-                    extension: Some(UdpFlowExtension { client_server_id, tunnel_id, port_offset }),
+                    extension: Some(UdpFlowExtension {
+                        client_server_id,
+                        tunnel_id,
+                        port_offset,
+                    }),
                 })
             }
             _ => Err(Some(footer_id)),
@@ -260,15 +309,30 @@ impl UdpFlow {
 
     pub fn footer_len(&self) -> usize {
         match self {
-            UdpFlow::V4 { extension: None, .. } => IP4_LEN_V1,
-            UdpFlow::V4 { extension: Some(_), frag: Some(_), .. } => IP4_LEN_V2_WITH_FRAG,
-            UdpFlow::V4 { extension: Some(_), frag: None, .. } => IP4_LEN_V2_WITHOUT_FRAG,
-            UdpFlow::V6 { extension: None, .. } => IP6_LEN_V1,
-            UdpFlow::V6 { extension: Some(_), .. } => IP6_LEN_V2,
+            UdpFlow::V4 {
+                extension: None, ..
+            } => IP4_LEN_V1,
+            UdpFlow::V4 {
+                extension: Some(_),
+                frag: Some(_),
+                ..
+            } => IP4_LEN_V2_WITH_FRAG,
+            UdpFlow::V4 {
+                extension: Some(_),
+                frag: None,
+                ..
+            } => IP4_LEN_V2_WITHOUT_FRAG,
+            UdpFlow::V6 {
+                extension: None, ..
+            } => IP6_LEN_V1,
+            UdpFlow::V6 {
+                extension: Some(_), ..
+            } => IP6_LEN_V2,
         }
     }
 
-    pub const MAX_IP4_LEN: usize = { m_max_list(&[IP4_LEN_V1, IP4_LEN_V2_WITH_FRAG, IP4_LEN_V2_WITHOUT_FRAG]) };
+    pub const MAX_IP4_LEN: usize =
+        { m_max_list(&[IP4_LEN_V1, IP4_LEN_V2_WITH_FRAG, IP4_LEN_V2_WITHOUT_FRAG]) };
 
     pub const MAX_IP6_LEN: usize = { m_max_list(&[IP6_LEN_V1, IP6_LEN_V2]) };
