+++ packages/playitd-tray/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "playitd-tray"
+version.workspace = true
+edition = "2024"
+authors = ["Patrick Lorio <patrick@playit.gg>"]
+description = "Windows tray application for managing the playit background service"
+repository = "https://github.com/playit-cloud/playit-agent"
+
+[dependencies]
+tokio = { workspace = true }
+playitd = { path = "../playitd", version = "0.17.1" }
+playit-ipc = { path = "../playit-ipc", version = "0.17.1" }
+
+[target."cfg(windows)".dependencies]
+image = { version = "0.25", default-features = false, features = ["ico"] }
+tray-icon = "0.22"
+windows-sys = { version = "0.59.0", features = [
+    "Win32_Foundation",
+    "Win32_Graphics_Gdi",
+    "Win32_System_Console",
+    "Win32_System_Com",
+    "Win32_System_LibraryLoader",
+    "Win32_Security",
+    "Win32_System_Services",
+    "Win32_System_Threading",
+    "Win32_UI_Shell",
+    "Win32_UI_WindowsAndMessaging",
+] }
